+++ providers/spotify.php
@@ -5,6 +5,7 @@ namespace Kirby\Plugins\distantnative\oEmbed\Providers;
 class Spotify extends Provider {
 
   protected function init() {
+    $this->getSizes();
     $this->getTheme();
     $this->getView();
   }
@@ -12,6 +13,7 @@ class Spotify extends Provider {
   public function code($code) {
     $this->setTheme();
     $this->setView();
+    $code = $this->setSizes($code);
     return $code;
   }
 
@@ -20,7 +22,7 @@ class Spotify extends Provider {
   // ================================================
 
   protected function getTheme() {
-    $this->theme = preg_match('/theme=(.+?)&?/U', $this->url, $t) ? $t[1] : false;
+    $this->theme = preg_match('/theme=([a-zA-Z]*)/', $this->url, $t) ? $t[1] : false;
   }
 
   protected function setTheme() {
@@ -29,12 +31,13 @@ class Spotify extends Provider {
     }
   }
 
+
   // ================================================
   //  View
   // ================================================
 
   protected function getView() {
-    $this->view = preg_match('/view=(.+?)&?/U', $this->url, $t) ? $t[1] : false;
+    $this->view = preg_match('/view=([a-zA-Z]*)/', $this->url, $t) ? $t[1] : false;
   }
 
   protected function setView() {
@@ -43,4 +46,23 @@ class Spotify extends Provider {
     }
   }
 
+
+  // ================================================
+  //  Sizes
+  // ================================================
+
+  protected function getSizes() {
+    $this->width = preg_match('/width=([0-9]*)/', $this->url, $t) ? $t[1] : false;
+    $this->height = preg_match('/height=([0-9]*)/', $this->url, $t) ? $t[1] : false;
+  }
+
+  protected function setSizes($code) {
+    if($this->width !== false) {
+      $code = str_ireplace('<iframe', '<iframe width="' . $this->width . '"', $code);
+    }
+    if($this->height !== false) {
+      $code = str_ireplace('<iframe', '<iframe height="' . $this->height . '"', $code);
+    }
+    return $code;
+  }
 }